Kce/linked list (#6)
Implements a linked list used by the Scalar_Expression_Node. It contains a list of terms. Also altered the Term type to have a fixed set of scalars based on the max dimension of the space. A little wasteful on dimensions less than the max but doesn't require a bunch of dynamic allocations when constructing a term.